### 2019-07-26 #1599 @whedon
[PRE REVIEW]: greta: simple and scalable statistical modelling in R
**Submitting author:** @goldingn (<a href="http://orcid.org/0000-0001-8916-5570">Nick Golding</a>)
**Repository:** <a href="https://github.com/greta-dev/greta" target ="_blank">https://github.com/greta-dev/greta</a>
**Version:** 0.3.0
**Reviewers:** @lionel68 @joethorley
