### 2019-07-26 #1599 @goldingn
The paper is deliberately a bit longer than most JOSS manuscripts - including a code example over two pages. I thought I might as well make it an interesting read. Happy to discuss cutting it down if that's an issue.
