### 2019-07-26 #1599 @goldingn
I suggested @usethedata for his statistical background. @terrytangyuan might also be interested in editing, given the TensorFlow angle.
