### 2019-07-27 #1599 @kyleniemeyer
@goldingn regarding the length, I don't think it is an issue. We don't publish articles with research reasults, but longer articles that describe the software and provide examples are fine (just not required).
