### 2019-07-27 #1601 @whedon
[REVIEW]: greta: simple and scalable statistical modelling in R
**Submitting author:** @goldingn (<a href="http://orcid.org/0000-0001-8916-5570">Nick Golding</a>)
**Repository:** <a href="https://github.com/greta-dev/greta" target ="_blank">https://github.com/greta-dev/greta</a>
**Version:** 0.3.1

@lionel68 @joethorley, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
