### 2019-07-30 greta-dev/greta#295 @joethorley
please add doi's to JOSS paper
For example the doi for

Carpenter, B., Gelman, A., Hoffman, M. D., Lee, D., Goodrich, B., Betancourt, M., Brubaker, M., et al. (2017). Stan: A probabilistic programming language. Journal of statistical software, 76 (1).

is 10.18637/jss.v076.i01

from https://www.jstatsoft.org/article/view/v076i01

for https://github.com/openjournals/joss-reviews/issues/1601#event-2514504784
