### 2019-07-30 greta-dev/greta#296 @joethorley
R code in JOSS paper appears to need manually wrapping
The third line of R code in the following snippet from page 2 is being cut off
```r
# prior
treatment_effects <- normal(1, 1, dim = 2, truncation = c(0, Inf))
post_treatment_rates <- treatment_effects[epil$trt_id] * baseline_rates[epil$subje...
```

https://github.com/openjournals/joss-reviews/issues/1601#event-2514504784
