### 2019-07-30 greta-dev/greta#297 @joethorley
some of the code fragments in the JOSS paper are missing new line separators
see

```r
distribution(baseline_y) <- poisson(baseline_rates * 8) distribution(epil$y) <- poisson(post_treatment_rates * 2)
```

vs

```r
subject_mean <- normal(0, 10)
subject_sd <- cauchy(0, 1, truncation = c(0, Inf))
```

https://github.com/openjournals/joss-reviews/issues/1601#event-2514504784
