### 2019-07-30 #1601 @joethorley
@goldingn greta is a fantastic piece of software. I just have three minor issues that need to be addressed

https://github.com/greta-dev/greta/issues/295
https://github.com/greta-dev/greta/issues/296
https://github.com/greta-dev/greta/issues/297
