### 2019-07-30 greta-dev/greta#297 @goldingn
Huh. I don't see this in [the pdf](https://github.com/openjournals/joss-papers/blob/joss.01601/joss.01601/10.21105.joss.01601.pdf). It seems to have a newline in the source and when I pandoc it locally.

Where are you seeing this issue?
