### 2019-07-30 016b69e
wrap code to fit in 80 characters; fixes #296
