+++ paper/paper.bib
@@ -1,8 +1,8 @@
 @article{openbugs,
-  title={OpenBUGS user manual, version 3.0. 2},
+  title={OpenBUGS user manual, version 3.2.3},
   author={Spiegelhalter, David and Thomas, Andrew and Best, Nicky and Lunn, Dave},
   journal={MRC Biostatistics Unit, Cambridge},
-  year={2007}
+  year={2014}
 }
 
 @inproceedings{jags,
@@ -33,14 +33,21 @@
   volume={76},
   number={1},
   year={2017},
-  publisher={Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA~…}
+  publisher={Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA~…},
+  doi = {10.18637/jss.v076.i01},
 }
 
 @article{pymc,
-  title={PyMC: Bayesian stochastic modelling in Python},
-  author={Fonnesbeck, Chris and Patil, Anand and Huard, David and Salvatier, John},
-  journal={Astrophysics Source Code Library},
-  year={2015}
+  doi = {10.7717/peerj-cs.55},
+  url = {https://doi.org/10.7717/peerj-cs.55},
+  year  = {2016},
+  month = {apr},
+  publisher = {{PeerJ}},
+  volume = {2},
+  pages = {e55},
+  author = {John Salvatier and Thomas V. Wiecki and Christopher Fonnesbeck},
+  title = {Probabilistic programming in Python using {PyMC}3},
+  journal = {{PeerJ} Computer Science}
 }
 
 @article{edward,
@@ -129,7 +136,7 @@
     url = {https://CRAN.R-project.org/package=bayesplot},
   }
 
-@Manual{r6,
+@manual{r6,
   title = {R6: Encapsulated Classes with Reference Semantics},
   author = {Winston Chang},
   year = {2019},
@@ -176,8 +183,7 @@
   url = {https://CRAN.R-project.org/package=tidyverse},
 }
 
-
-@Manual{Rcore,
+@manual{Rcore,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
