### 2019-07-31 greta-dev/greta#295 @goldingn
OK, I think all the papers with dois have their dois in the citations now
