### 2019-07-31 greta-dev/greta#295 @goldingn
(resolved in 3f2ba9027)
