### 2019-07-31 #1601 @goldingn
Thanks @joethorley!

Resolved two of those issues now. The last formatting issue (#297) doesn't seem to be present in the article proof, but I may have misunderstood.
