### 2019-07-31 #1601 @terrytangyuan
Nice job @goldingn and thanks @joethorley for the review!
👋 @lionel68 Have you got a chance to review this paper yet?
