### 2019-07-31 greta-dev/greta#297 @joethorley
Sorry I explained it poorly - it only appears when you try and cut and paste the text into an R console
