### 2019-07-31 greta-dev/greta#297 @goldingn
Aah, yeah that makes sense.

I can replicate this with a local build of the pdf. The newline separators are there in the Rmd, md, and tex files, but when I copy-paste from the (local) pdf the line breaks disappear for some lines (I can't spot any obvious pattern for when they disappear). I think that's not unexpected for pdf - it's not really designed to be copy-pasted from.

However I can't replicate the issue with [the JOSS pdf](https://github.com/openjournals/joss-papers/blob/joss.01601/joss.01601/10.21105.joss.01601.pdf). All the newlines are for me when copy-pasting from that one. From (hastily) reading [some stackoverflow messages](https://tex.stackexchange.com/questions/367364/copying-continuous-text-from-pdf-file-to-txt-file) and looking at the whedon build script, it seems like [the JOSS latex template](https://github.com/openjournals/whedon/blob/master/resources/latex.template#L182-L184) uses some options that solve this issue.
