### 2019-08-02 greta-dev/greta#299 @lionel68
Could not run example model in JOSS paper
I tried to run the example model in the JOSS paper in my R session but got the following error:

```r
Error in dim(free_state_draws) <- c(1, dim(free_state_draws)) :
  invalid first argument
```
Here is my sessionInfo():

```r
R version 3.6.0 (2019-04-26)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 18.04.2 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/openblas/libblas.so.3
LAPACK: /usr/lib/x86_64-linux-gnu/libopenblasp-r0.2.20.so

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=de_BE.UTF-8        LC_COLLATE=en_US.UTF-8     LC_MONETARY=de_BE.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=de_BE.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=de_BE.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] greta_0.3.0   MASS_7.3-51.1

 [1] Rcpp_1.0.1        rstudioapi_0.10   whisker_0.3-2     magrittr_1.5      hms_0.4.2         progress_1.2.0    lattice_0.20-38   R6_2.4.0          rlang_0.3.4       globals_0.12.4
[11] tools_3.6.0       parallel_3.6.0    grid_3.6.0        coda_0.19-2       tfruns_1.4        digest_0.6.18     assertthat_0.2.0  crayon_1.3.4      tensorflow_1.14.0 Matrix_1.2-17
[21] base64enc_0.1-3   codetools_0.2-16  compiler_3.6.0    prettyunits_1.0.2 reticulate_1.13   jsonlite_1.6      future_1.13.0     listenv_0.7.0     pkgconfig_2.0.2
```
And the code I used:

```r
library(MASS)
library(greta)

epil$trt_id <- as.numeric(epil$trt)
baseline_y <- epil$base[!duplicated(epil$subject)]
subject_mean <- normal(0, 10)
subject_sd <- cauchy(0, 1, truncation = c(0, Inf))
subject_effects <- normal(subject_mean, subject_sd, dim = 59)
subject_sd <- cauchy(0, 1, truncation = c(0, Inf))
subject_effects <- normal(subject_mean, subject_sd, dim = 59)
baseline_rates <- exp(subject_effects)
treatment_effects <- normal(1, 1, dim = 2, truncation = c(0, Inf))
post_treatment_rates <- treatment_effects[epil$trt_id] * baseline_rates[epil$subject]
distribution(baseline_y) <- poisson(baseline_rates * 8)
distribution(epil$y) <- poisson(post_treatment_rates * 2)
m <- model(treatment_effects, subject_sd)
draws <- mcmc(m, chains = 3)
```
