### 2019-08-02 greta-dev/greta#300 @lionel68
Python warnings when running greta::model 
I get python warnings when I run "model" for the first time in an R session:

```r
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/dtypes.py:516: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_qint8 = np.dtype([("qint8", np.int8, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/dtypes.py:517: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_quint8 = np.dtype([("quint8", np.uint8, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/dtypes.py:518: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_qint16 = np.dtype([("qint16", np.int16, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/dtypes.py:519: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_quint16 = np.dtype([("quint16", np.uint16, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/dtypes.py:520: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_qint32 = np.dtype([("qint32", np.int32, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/dtypes.py:525: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  np_resource = np.dtype([("resource", np.ubyte, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorboard/compat/tensorflow_stub/dtypes.py:541: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_qint8 = np.dtype([("qint8", np.int8, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorboard/compat/tensorflow_stub/dtypes.py:542: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_quint8 = np.dtype([("quint8", np.uint8, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorboard/compat/tensorflow_stub/dtypes.py:543: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_qint16 = np.dtype([("qint16", np.int16, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorboard/compat/tensorflow_stub/dtypes.py:544: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_quint16 = np.dtype([("quint16", np.uint16, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorboard/compat/tensorflow_stub/dtypes.py:545: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  _np_qint32 = np.dtype([("qint32", np.int32, 1)])
/home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorboard/compat/tensorflow_stub/dtypes.py:550: FutureWarning: Passing (type, 1) or '1type' as a synonym of type is deprecated; in a future version of numpy, it will be understood as (type, (1,)) / '(1,)type'.
  np_resource = np.dtype([("resource", np.ubyte, 1)])
WARNING: Logging before flag parsing goes to stderr.
W0802 10:21:47.772652 139702542626752 deprecation_wrapper.py:119] From /home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/util/deprecation_wrapper.py:119: The name tf.placeholder is deprecated. Please use tf.compat.v1.placeholder instead.

W0802 10:21:48.031610 139702542626752 deprecation_wrapper.py:119] From /home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/util/deprecation_wrapper.py:119: The name tf.ConfigProto is deprecated. Please use tf.compat.v1.ConfigProto instead.

W0802 10:21:48.036754 139702542626752 deprecation_wrapper.py:119] From /home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/util/deprecation_wrapper.py:119: The name tf.OptimizerOptions is deprecated. Please use tf.compat.v1.OptimizerOptions instead.

W0802 10:21:48.041310 139702542626752 deprecation_wrapper.py:119] From /home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/util/deprecation_wrapper.py:119: The name tf.Session is deprecated. Please use tf.compat.v1.Session instead.

W0802 10:21:48.078707 139702542626752 deprecation_wrapper.py:119] From /home/lionel/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/util/deprecation_wrapper.py:119: The name tf.global_variables_initializer is deprecated. Please use tf.compat.v1.global_variables_initializer instead.
```

sessionInfo identical to #299
