### 2019-08-02 greta-dev/greta#301 @lionel68
Textual comments for the JOSS paper
I have a few minor textual comments:

* "greta is **a** package for ..."
* insert line break to keep __greta\_array__ together: "by creating and manipulating [line break] greta_array"
* sentence unclear: "information on the greta arrays from which they were created or which re created them"
