### 2019-08-02 #1601 @lionel68
greta is a very promising software and is likely to become a very handy tool in the data scientist toolbox.

I opened three issues:

greta-dev/greta#299 (example R code not running on my laptop)
greta-dev/greta#300 (python warnings when running greta::model)
greta-dev/greta#301 (minor textual comments)

Looking forward to using the capacities of greta.
