### 2019-08-02 #1601 @kyleniemeyer
@lionel68 can you edit the checklist now?
