### 2019-08-05 greta-dev/greta#299 @chameroy
I'm receiving the same error just by trying the example model under the "get started" section.

I'm running R version 3.6.1 on Windows 10
