### 2019-08-06 greta-dev/greta#299 @goldingn
Thanks @lionel68.

I think this is due to you using a recent version of TF/TFP, but the CRAN (rather than dev) version of greta (0.3.0). If you install the most recent version of greta (0.3.0.9008), from GitHub I think it should work:

```r
remotes::install_github("greta-dev/greta")
```

Up until now, I've been specifying minimum versions of TF and TFP as dependencies (ie. TF >= 1.10; TFP >=0.3.0). I think that's generally considered better practice than depending on specific versions. But the APIs for both of these packages (but particularly TFP) change so frequently, that errors keep occurring, making it impossible to maintain. So the development (GitHub) version of greta now depends on specific releases (TF = 1.14, TFP = 0.7.0), which is what I'm planning to do in the future.

I can see why you went for the CRAN version, and the JOSS review is for greta 0.3.0, but I'm not clear on how that works with dev versions. I'll ask the editor's opinion over on that channel.
