### 2019-08-06 greta-dev/greta#300 @goldingn
Would you mind checking for these errors with the latest greta (as in #299), and posting the output if they are different in the new one?
