### 2019-08-06 greta-dev/greta#301 @goldingn
Thanks! Will make these edits tomorrow.
