### 2019-08-06 #1601 @goldingn
Hi @terrytangyuan, how does the JOSS review process deal with changing version numbers, and things like CRAN submission?

There's a version 0.3.0 of greta on CRAN (and in the GitHub commit history), and a development version currently at 0.3.0.9008 on GitHub. On submission I stated that review is for 0.3.0, and @lionel68 reviewed that version exactly. He hit an issue which I'm pretty sure is fixed in the dev version.

I'd like to wrap that and any other fixes that come from this review into a 0.3.1 release candidate and push that to CRAN. I'd like all non-dev versions to be released on CRAN.

How does the following procedure sound:

1. complete review of the paper and software on the dev version of greta
2. submit that to CRAN as 0.3.1
3. on acceptance to CRAN, update both the greta master branch and the JOSS version of record to 0.3.1
4. hit the publish button on JOSS

I'm sure this is something that has come up before for the journal, so let me know if I'm overthinking it.

---

P.S. I'd also love to have the JOSS citation in the next CRAN release, but that might be a bit circular!
