### 2019-08-06 #1601 @terrytangyuan
@goldingn Yes, we can update the version number here through a `@whedon` command prior to JOSS paper publication/acceptance. So your outlined procedure looks fine to me. Before step 4, you'll need to create an archive on services like Zenondo (I'll provide more details on the requirements once accepted).
