### 2019-08-06 #1601 @goldingn
Great, thanks @terrytangyuan. I'll do that and leave off on the citation then.
