### 2019-08-07 greta-dev/greta#300 @lionel68
These warnings disappeared in the latest dev greta version
