### 2019-08-07 #1601 @lionel68
The issue I had with the example code were resolved by installing the dev version of greta from Github. I now checked all review items in the list above.
