### 2019-08-07 dd6ade0
fix @lionel68's textuals comments (closes #301)
