+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: greta
 Type: Package
 Title: Simple and Scalable Statistical Modelling in R
-Version: 0.3.0.9009
+Version: 0.3.0.9010
 Date: 2019-06-03
 Authors@R: c(
   person("Nick", "Golding", role = c("aut", "cre"),
