### 2019-08-07 d9e0430
switch to pure md readme and edit
