+++ README.md
@@ -1,6 +1,6 @@
-![](README_files/top_banner.png)
+![](logos/top_banner.png)
 
-### greta is an R package for writing statistical models and fitting them by MCMC.
+### greta is an R package for writing statistical models and fitting them by MCMC and optimisation
 
 greta lets you write your own model like in BUGS, JAGS and Stan, except
 that you write models right in R, it scales well to massive datasets,
@@ -8,7 +8,7 @@ and it’s easy to extend and build on.
 
 ### See the [website](https://greta-stats.org/) for more information, including [tutorials](https://greta-stats.org/articles/get_started.html), [examples](https://greta-stats.org/articles/example_models.html), [package documentation](https://greta-stats.org/reference/index.html), and the [greta forum](https://forum.greta-stats.org).
 
-You can install the current release version of the package (0.3.0) from
+You can install the current release version of the package from
 CRAN:
 
 ``` r
@@ -28,4 +28,5 @@ status](https://travis-ci.org/greta-dev/greta.svg?branch=master)](https://travis
 version](http://www.r-pkg.org/badges/version/greta)](https://cran.rstudio.com/web/packages/greta)
 [![license](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 [![doi](https://zenodo.org/badge/73758247.svg)](https://zenodo.org/badge/latestdoi/73758247)
-![](README_files/bottom_banner.png)
+[![joss](http://joss.theoj.org/papers/62e62338d8ba44ff9e5dc295cf631477/status.svg)](http://joss.theoj.org/papers/62e62338d8ba44ff9e5dc295cf631477)
+![](logos/bottom_banner.png)
