### 2019-08-08 edf268b
run long travis checks
