### 2019-08-08 e65e70a
update NEWS.md
