### 2019-08-08 b6beae4
try to make travis wait 30 minutes for tests
