### 2019-08-08 fbc808d
fix dead links in package-level helpfile
