+++ R/package.R
@@ -12,10 +12,10 @@
 #'   clusters or on GPUs.
 #'
 #'   See the simple example below, and take a look at the
-#'   \href{https://greta-dev.github.io/greta}{greta website} for more information
+#'   \href{https://greta-stats.org}{greta website} for more information
 #'   including
-#'   \href{https://greta-dev.github.io/greta/get_started.html}{tutorials} and
-#'   \href{https://greta-dev.github.io/greta/example_models.html}{examples}.
+#'   \href{https://greta-stats.org/articles/get_started.html}{tutorials} and
+#'   \href{https://greta-stats.org/articles/example_models.html}{examples}.
 #'
 #' @docType package
 #' @importFrom tensorflow tf
