### 2019-08-08 fd858eb
improve warmup for rwmh in Geweke tests
