### 2019-08-08 0c452fe
actually *lower* the sensitivity of the geweke test to false positives this time
