### 2019-08-08 ac34041
give up on doing longer release checks on travis
