### 2019-08-08 #1601 @goldingn
The new release is running the gauntlet at CRAN...
