### 2019-08-08 c631e3d
edits to Description as requested by CRAN
