+++ DESCRIPTION
@@ -19,7 +19,9 @@ Authors@R: c(
   person("Paul", "Teetor", role = "ctb"),
   person("Jian", "Yen", role = "ctb")
   )
-Description: Write statistical models in R and fit them by MCMC on CPUs and GPUs, using Google TensorFlow.
+Description: Write statistical models in R and fit them by MCMC and optimisation on CPUs and GPUs, using Google 'TensorFlow'.
+  greta lets you write your own model like in BUGS, JAGS and Stan, except that you write models right in R, it scales well to massive datasets, and it’s easy to extend and build on.
+  See the website for more information, including tutorials, examples, package documentation, and the greta forum.
 License: Apache License 2.0
 URL: https://greta-stats.org
 BugReports: https://github.com/greta-dev/greta/issues
