### 2019-08-08 0d258a3
use absolute weblinks (not relative links) in vignette
