+++ NEWS.md
@@ -1,3 +1,5 @@
+# greta (development version)
+
 # greta 0.3.1
 
 This release is predominantly a patch to make greta work with recent versions of TensorFlow and TensorFlow Probability, which were not backward compatible with the versions on which greta previously depended. From this release forward, greta will depend on specific (rather than minimum) versions of these two pieces of software to avoid it breaking if more changes are made to the APIS of these packages.
