+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: greta
 Type: Package
 Title: Simple and Scalable Statistical Modelling in R
-Version: 0.3.1
+Version: 0.3.1.9000
 Date: 2019-08-08
 Authors@R: c(
   person("Nick", "Golding", role = c("aut", "cre"),
