### 2019-08-09 #1601 @goldingn
OK, accepted on CRAN and updated on Github master.

@terrytangyuan, could you please change the version of record to 0.3.1?

I think all of the reviewers' comments have been addressed now too.
