### 2019-08-09 #1601 @terrytangyuan
Thank you!

* The title of the archive is the same as the JOSS paper title
* That the authors of the archive are the same as the JOSS paper authors
