### 2019-08-10 #1601 @goldingn
This version of the software is now released on GitHub and on CRAN. Here's a figshare archive of this version of the software:
https://doi.org/10.6084/m9.figshare.9461447.v1
Title and authors are the same as this paper.
