+++ tests/testthat/test_functions.R
@@ -75,21 +75,25 @@ test_that("matrix functions work as expected", {
 })
 
 test_that("kronecker works with greta and base array arguments", {
-    a <- rWishart(1, 6, diag(5))[, , 1]
-    b <- chol(a)
 
-    a_greta <- as_data(a)
-    b_greta <- as_data(b)
+  skip_if_not(check_tf_version())
+  source("helpers.R")
 
-    base_out <- kronecker(a, b)
-    greta_out1 <- kronecker(a_greta, b)
-    greta_out2 <- kronecker(a, b_greta)
+  a <- rWishart(1, 6, diag(5))[, , 1]
+  b <- chol(a)
 
-    expect_true(is.greta_array(greta_out1))
-    expect_true(is.greta_array(greta_out1))
+  a_greta <- as_data(a)
+  b_greta <- as_data(b)
 
-    compare_op(base_out, grab(greta_out1))
-    compare_op(base_out, grab(greta_out2))
+  base_out <- kronecker(a, b)
+  greta_out1 <- kronecker(a_greta, b)
+  greta_out2 <- kronecker(a, b_greta)
+
+  expect_true(is.greta_array(greta_out1))
+  expect_true(is.greta_array(greta_out1))
+
+  compare_op(base_out, grab(greta_out1))
+  compare_op(base_out, grab(greta_out2))
 
 })
 
