### 2019-08-10 27b1ed3
deprecate irreplaceable SciPy optimisers
