### 2019-08-10 e1b1298
style edits from linter
