+++ R/optimisers.R
@@ -81,43 +81,66 @@ define_tf_optimiser <- function(name,
 #' @export
 #'
 nelder_mead <- function() {
-  define_scipy_optimiser("nelder_mead",
-                         method = "Nelder-Mead")
+
+  define_scipy_optimiser(
+    name = "nelder_mead",
+    method = "Nelder-Mead"
+  )
+
 }
 
 #' @rdname optimisers
 #' @export
 #'
 powell <- function() {
+
   optimiser_deprecation_warning()
-  define_scipy_optimiser("powell",
-                         method = "Powell")
+
+  define_scipy_optimiser(
+    name = "powell",
+    method = "Powell"
+  )
+
 }
 
 #' @rdname optimisers
 #' @export
 #'
 cg <- function() {
+
   optimiser_deprecation_warning()
-  define_scipy_optimiser("cg",
-                         method = "CG")
+
+  define_scipy_optimiser(
+    name = "cg",
+    method = "CG"
+  )
+
 }
 
 #' @rdname optimisers
 #' @export
 #'
 bfgs <- function() {
-  define_scipy_optimiser("bfgs",
-                         method = "BFGS")
+
+  define_scipy_optimiser(
+    name = "bfgs",
+    method = "BFGS"
+  )
+
 }
 
 #' @rdname optimisers
 #' @export
 #'
 newton_cg <- function() {
+
   optimiser_deprecation_warning()
-  define_scipy_optimiser("newton_cg",
-                         method = "Newton-CG")
+
+  define_scipy_optimiser(
+    name = "newton_cg",
+    method = "Newton-CG"
+  )
+
 }
 
 #' @rdname optimisers
@@ -128,13 +151,18 @@ newton_cg <- function() {
 #' @param maxls maximum number of line search steps per iteration
 #'
 l_bfgs_b <- function(maxcor = 10, maxls = 20) {
+
   optimiser_deprecation_warning()
-  define_scipy_optimiser("l_bfgs_b",
-                         method = "L-BFGS-B",
-                         parameters = list(
-                           maxcor = as.integer(maxcor),
-                           maxls = as.integer(maxls)
-                         ))
+
+  define_scipy_optimiser(
+    name = "l_bfgs_b",
+    method = "L-BFGS-B",
+    parameters = list(
+      maxcor = as.integer(maxcor),
+      maxls = as.integer(maxls)
+    )
+  )
+
 }
 
 #' @rdname optimisers
@@ -145,14 +173,19 @@ l_bfgs_b <- function(maxcor = 10, maxls = 20) {
 #' @param rescale log10 scaling factor used to trigger rescaling of objective
 #'
 tnc <- function(max_cg_it = -1, stepmx = 0, rescale = -1) {
+
   optimiser_deprecation_warning()
-  define_scipy_optimiser("tnc",
-                         method = "TNC",
-                         parameters = list(
-                           maxCGit = as.integer(max_cg_it),
-                           stepmx = stepmx,
-                           rescale = rescale
-                         ))
+
+  define_scipy_optimiser(
+    name = "tnc",
+    method = "TNC",
+    parameters = list(
+      maxCGit = as.integer(max_cg_it),
+      stepmx = stepmx,
+      rescale = rescale
+    )
+  )
+
 }
 
 #' @rdname optimisers
@@ -164,22 +197,32 @@ tnc <- function(max_cg_it = -1, stepmx = 0, rescale = -1) {
 #'   iterations nor convergence, so these elements of the output are set to NA
 #'
 cobyla <- function(rhobeg = 1) {
+
   optimiser_deprecation_warning()
-  define_scipy_optimiser("cobyla",
-                         method = "COBYLA",
-                         parameters = list(
[…]
