+++ tests/spelling.R
@@ -1,4 +1,4 @@
-if (requireNamespace('spelling', quietly = TRUE)) {
+if (requireNamespace("spelling", quietly = TRUE)) {
 
   spelling::spell_check_test(
     vignettes = TRUE,
