### 2019-08-10 cfcdc05
suppress future multicore warning during tests
