### 2019-08-10 f85d6a2
more linting exclusions
