### 2019-08-10 6c78e3d
use CRAN DiagrammerSVG on travis
