### 2019-08-10 87b29cb
suppress TensorFlow deprecation warnings
