+++ R/package.R
@@ -47,6 +47,9 @@ tfp <- reticulate::import("tensorflow_probability", delay_load = TRUE)
   # silence TF's CPU instructions message
   Sys.setenv(TF_CPP_MIN_LOG_LEVEL = 2)
 
+  # silence messages about deprecation etc.
+  disable_tensorflow_logging()
+
   # warn if TF version is bad
   check_tf_version("startup")
 
