+++ NAMESPACE
@@ -217,6 +217,7 @@ importFrom(grDevices,col2rgb)
 importFrom(grDevices,colorRampPalette)
 importFrom(progress,progress_bar)
 importFrom(reticulate,conda_binary)
+importFrom(reticulate,import)
 importFrom(reticulate,py_available)
 importFrom(reticulate,py_module_available)
 importFrom(reticulate,py_set_attr)
