### 2019-08-10 a93d71c
remove redundant function
