### 2019-08-10 9c84cf8
avoid lintr false positive; re-animate the lintr bot
