+++ .travis.yml
@@ -5,7 +5,7 @@ dist: trusty
 env:
   global:
   - RELEASE_CANDIDATE=false
-  - LINTR_COMMENT_BOT=false
+  - LINTR_COMMENT_BOT=true
   - secure: Q34zwlFiE+zShcPgTM99NILILBG2ViTaVo1ufRB/GFniGbEBIOwliqnozzWhhbuHEHG64qCNGbM4pMhyekC5YC9Rwp0QEt7sUk/tpV/17tkDiwjfEICAWsXK9rLrP3d3Pe8R/PYd4vyh8DiwgRl8woHCmmMJlIU9D/4XSOF6a325LX7U1ubJP5p5FrAdHt1ki4vEdUti4b7CdHZJHPjger+C+TNvvb75cIXR4RBbAKBHmpg/rTBASMRthJcKOGfym7QDsSrynQqMKR/BJsqzWvGzGYlUtGq04tHBbE9AVmcgRMqA5UuBcwmrKwQarx/yEX5YG8pVG2IgWKxowuxf+DGSwQ84m3D58fmiDcvdfNV5Spr9/lzXXMu3ae0WBbLOSIAc3FPJaEW4NK/HkeXrqbFnDy1dm0trLvG5AE0XisTcwFpoyx8+PnbAAa9fkJSOP+1n8472ym6OGRfngs1GoMrFaCpUaLug9eux2Ckc1sq2cLOo9wWzTsIoFDKCS5uBEvj1Vw1I845iO0fymj0IGTpLtjoNHGNMVyb+YQ+yTIIjW/8sgYhj+NaDqvjCwSgaHEJbWouCgoxsFhKaXsivosWZDebMM1PLsUaDyShv3DwOuFv5xD+0pX4AiuSDhhrVO7yThgiqD9KCa/30d1/CXU/GSTV5OlShJLz2XsLr7+E=
 
 r_binary_packages:
