### 2019-08-11 0592b19
make spellcheck happen outside tests
