+++ tests/spelling.R
@@ -1,9 +1,0 @@
-if (requireNamespace("spelling", quietly = TRUE)) {
-
-  spelling::spell_check_test(
-    vignettes = TRUE,
-    error = FALSE,
-    skip_on_cran = TRUE
-  )
-
-}
