### 2019-08-11 64ebac4
remove some spell check false negatives from WORDLIST and fix in text
