### 2019-08-11 6086b9f
remove some spell check false negatives from WORDLIST and fix in text
