+++ man/distributions.Rd
@@ -149,7 +149,7 @@ The discrete probability distributions (\code{bernoulli},
   corresponds to an independent realisation. If a single realisation or
   parameter value is specified, it must therefore be a row vector (see
   example). \code{n_realisations} gives the number of rows/realisations, and
-  \code{dimension} gives the dimension of the distribution. Ie. a bivariate
+  \code{dimension} gives the dimension of the distribution. I.e. a bivariate
   normal distribution would be produced with \code{multivariate_normal(...,
   dimension = 2)}. The dimension can usually be detected from the parameters.
 
